Execute command to kill process called killmenow

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,8 @@
-# Manifest tat kills process killmenow
+# Manifest that kills process killmenow
 exec { 'killmenow':
 command     => 'pkill killmenow',
+onlyif      => 'pgrep killmenow',
+provider    => 'shell',
+creates     => '/killmenow',
 path        => '/usr/bin',
-refreshonly => true,
 }
